fix for weird reloading behavior
