+++ frontend/plan/server.js
@@ -28,14 +28,18 @@ app.prepare()
     .then(() => {
         server = express();
 
-        // Set up the proxy.
-        if (dev && devProxy) {
-            // eslint-disable-next-line
-            const { createProxyMiddleware } = require("http-proxy-middleware");
-            // eslint-disable-next-line
-            Object.keys(devProxy).forEach(function (context) {
-                server.use(createProxyMiddleware(context, devProxy[context]));
-            });
+        if (dev) {
+            server.on('upgrade', handle)
+
+            // Set up the proxy.
+            if (devProxy) {
+                // eslint-disable-next-line
+                const { createProxyMiddleware } = require("http-proxy-middleware");
+                // eslint-disable-next-line
+                Object.keys(devProxy).forEach(function (context) {
+                    server.use(createProxyMiddleware(context, devProxy[context]));
+                });
+            }
         }
 
         // Default catch-all handler to allow Next.js to handle all other routes
